Add wasm_utils & refactor prelude module
