+++ src/main.rs
@@ -1,9 +1,11 @@
-use xkpasswd::prelude::*;
-use xkpasswd::settings::*;
+mod prelude;
+pub mod settings;
+
+use prelude::*;
+use settings::*;
 
 fn main() {
-    let dict_en_bytes = include_bytes!("./assets/dict_en.txt");
-    let dict = &load_dict(&dict_en_bytes[..]);
+    let pass_generator = Xkpasswd::new();
     let settings = &Settings::default().words_count(3).word_lengths(5, 8);
-    println!("{}", gen_passwd(dict, settings));
+    println!("{}", pass_generator.gen_pass(settings));
 }
