+++ src/wasm_utils.rs
@@ -1,0 +1,14 @@
+extern crate web_sys;
+
+macro_rules! console_log {
+    ( $( $t:tt )* ) => {
+        web_sys::console::log_1(&format!( $( $t )* ).into());
+    }
+}
+
+pub(crate) use console_log;
+
+pub fn set_panic_hook() {
+    #[cfg(feature = "console_error_panic_hook")]
+    console_error_panic_hook::set_once();
+}
